+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.2.0] - 2025-09-29
+### Added
+- [#07] Configure annotations for loadbalancer via values.yaml
+
 ## [v0.0.1] - 2025-09-04
 ### Added
 - [#01] Provide helm chart to install core components of the ecosystem
+- [#05] post-install job to initialize default configuration